Found some errors in serverConnection. Test after correct server.

+++ app/src/main/java/ru/project/iidea/ResponseViewFragmentInterface.java
@@ -2,5 +2,5 @@ package ru.project.iidea;
 
 public interface ResponseViewFragmentInterface {
     public void onBackButtonPressed();
-    public void onUserIdClicked(long userID);
+    public void onUserIdClicked(int userID);
 }
